Move common-tools to utils package
