+++ src/components/panel-body/UESpage/component.js
@@ -1,10 +1,6 @@
 import { Grid } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
-import {
-  applyFilters,
-  sortOnColumnCompareFunction,
-  updateStateWithDates,
-} from 'common-tools/functions';
+import { applyFilters, sortOnColumnCompareFunction, updateStateWithDates } from 'utils/functions';
 import surveyUnitDBService from 'indexedbb/services/surveyUnit-idb-service';
 import PropTypes from 'prop-types';
 import React, { useEffect, useState } from 'react';
