+++ src/components/panel-body/UEpage/contacts/contactAttempts/component.js
@@ -1,6 +1,6 @@
 import { makeStyles, Paper, Typography } from '@material-ui/core';
-import formEnum from 'common-tools/enum/formEnum';
-import { getSortedContactAttempts } from 'common-tools/functions';
+import formEnum from 'utils/enum/formEnum';
+import { getSortedContactAttempts } from 'utils/functions';
 import D from 'i18n';
 import PropTypes from 'prop-types';
 import React, { useContext, useEffect, useState } from 'react';
