+++ src/components/panel-body/UEpage/details/contact.js
@@ -1,6 +1,6 @@
 import { Grid } from '@material-ui/core';
-import formEnum from 'common-tools/enum/formEnum';
-import { getMailData, getPhoneData, getUserData } from 'common-tools/functions';
+import formEnum from 'utils/enum/formEnum';
+import { getMailData, getPhoneData, getUserData } from 'utils/functions';
 import D from 'i18n';
 import PropTypes from 'prop-types';
 import React from 'react';
