+++ src/components/panel-body/UEpage/infoTile.js
@@ -4,7 +4,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import PersonIcon from '@material-ui/icons/Person';
 import { Skeleton } from '@material-ui/lab';
-import { getprivilegedPerson } from 'common-tools/functions';
+import { getprivilegedPerson } from 'utils/functions';
 import surveyUnitDBService from 'indexedbb/services/surveyUnit-idb-service';
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
