+++ src/components/panel-body/UEpage/stateLine.js
@@ -4,8 +4,8 @@ import Stepper from '@material-ui/core/Stepper';
 import { makeStyles } from '@material-ui/core/styles';
 import CheckCircleOutlineIcon from '@material-ui/icons/CheckCircleOutline';
 import RadioButtonUncheckedIcon from '@material-ui/icons/RadioButtonUnchecked';
-import toDoEnum from 'common-tools/enum/SUToDoEnum';
-import { convertSUStateInToDo, getLastState } from 'common-tools/functions';
+import toDoEnum from 'utils/enum/SUToDoEnum';
+import { convertSUStateInToDo, getLastState } from 'utils/functions';
 import PropTypes from 'prop-types';
 import React, { useContext } from 'react';
 import SurveyUnitContext from './UEContext';
