+++ src/utils/api/surveyUnitAPI.js
@@ -10,11 +10,13 @@ export const getSurveyUnits = (urlPearApi, authenticationMode) =>
         })
           .then(res => resolve(res))
           .catch(e => {
-            reject(new Error(`Failed to fetch survey-units : ${e.response.data.error.message}`));
+            reject(
+              new Error(`Failed to fetch survey-units : ${e?.response?.data?.error?.message}`)
+            );
           });
       })
       .catch(e => {
-        reject(new Error(`Error during refreshToken : ${e.response.data.error.message}`));
+        reject(new Error(`Error during refreshToken : ${e?.response?.data?.error?.message}`));
       });
   });
 
@@ -26,9 +28,9 @@ export const getSurveyUnitById = (urlPearApi, authenticationMode) => id =>
           headers: getHeader(authenticationMode),
         })
           .then(res => resolve(res))
-          .catch(e => reject(new Error(`Failed to fetch survey-unit (id:${id}) : ${e.message}`)));
+          .catch(e => reject(new Error(`Failed to fetch survey-unit (id:${id}) : ${e?.message}`)));
       })
-      .catch(e => reject(new Error(`Error during refreshToken : ${e.message}`)));
+      .catch(e => reject(new Error(`Error during refreshToken : ${e?.message}`)));
   });
 
 export const putDataSurveyUnitById = (urlPearApi, authenticationMode) => (id, su) =>
@@ -39,7 +41,7 @@ export const putDataSurveyUnitById = (urlPearApi, authenticationMode) => (id, su
           headers: getHeader(authenticationMode),
         })
           .then(res => resolve(res))
-          .catch(e => reject(new Error(`Failed to put survey-unit (id:${id}) : ${e.message}`)));
+          .catch(e => reject(new Error(`Failed to put survey-unit (id:${id}) : ${e?.message}`)));
       })
-      .catch(e => reject(new Error(`Error during refreshToken : ${e.message}`)));
+      .catch(e => reject(new Error(`Error during refreshToken : ${e?.message}`)));
   });
