+++ src/utils/icons/SyncIcon.js
